Using non-TLS connections in tests
