+++ internal/net/client_grpc.go
@@ -2,7 +2,6 @@ package net
 
 import (
 	"context"
-	"crypto/tls"
 	"errors"
 	"fmt"
 	"io"
@@ -10,16 +9,12 @@ import (
 	"sync"
 	"time"
 
-	"go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc"
 	"golang.org/x/net/proxy"
 	"google.golang.org/grpc"
-	"google.golang.org/grpc/connectivity"
-	"google.golang.org/grpc/credentials"
 	"google.golang.org/grpc/encoding/gzip"
 
 	"github.com/drand/drand/v2/common/log"
 	"github.com/drand/drand/v2/common/tracer"
-	"github.com/drand/drand/v2/internal/metrics"
 	"github.com/drand/drand/v2/protobuf/drand"
 )
 
@@ -228,51 +223,6 @@ func (g *grpcClient) Status(ctx context.Context, p Peer, in *drand.StatusRequest
 	defer cancel()
 	resp, err = client.Status(ctx, in, opts...)
 	return resp, err
-}
-
-// conn retrieve an already existing conn to the given peer or create a new one
-func (g *grpcClient) conn(ctx context.Context, p Peer) (*grpc.ClientConn, error) {
-	g.Lock()
-	defer g.Unlock()
-	var err error
-
-	// we try to retrieve an existing connection if available
-	c, ok := g.conns[p.Address()]
-	if ok && (c.GetState() == connectivity.Shutdown || c.GetState() == connectivity.TransientFailure) {
-		ok = false
-		delete(g.conns, p.Address())
-		metrics.OutgoingConnectionState.WithLabelValues(p.Address()).Set(float64(c.GetState()))
-	}
-
-	// otherwise we try to re-dial it
-	if !ok {
-		g.log.Debugw("initiating new grpc conn using TLS", "to", p.Address())
-		var opts []grpc.DialOption
-
-		config := &tls.Config{MinVersion: tls.VersionTLS12}
-		opts = append(opts, g.opts...)
-		opts = append(opts,
-			grpc.WithStatsHandler(otelgrpc.NewClientHandler()),
-		)
-
-		c, err = grpc.DialContext(ctx, p.Address(), append(opts,
-			grpc.WithTransportCredentials(credentials.NewTLS(config)))...)
-		if err != nil {
-			g.log.Errorw("error initiating a new grpc conn using TLS", "to", p.Address(), "err", err)
-			// We increase the GroupDialFailures counter when both failed
-			metrics.GroupDialFailures.WithLabelValues(p.Address()).Inc()
-		} else {
-			g.log.Debugw("new grpc conn established", "state", c.GetState(), "to", p.Address())
-			g.conns[p.Address()] = c
-			metrics.OutgoingConnections.Set(float64(len(g.conns)))
-		}
-	}
-
-	// Emit the connection state regardless of whether it's a new or an existing connection
-	if err == nil {
-		metrics.OutgoingConnectionState.WithLabelValues(p.Address()).Set(float64(c.GetState()))
-	}
-	return c, err
 }
 
 func (g *grpcClient) Stop() {
